requirements.txt
sqlalchemy==2.0.19
pandas==2.0.3
psycopg2-binary==2.9.7
pymysql==1.1.0
python-dotenv==1.0.0